Fix String bridging tests in use_os_stdlib mode
